Move discord specific interface implementations to their respective model directory
