+++ src/store/commands/command-execution-data.ts
@@ -1,4 +1,4 @@
-import Message from "@model/message";
+import Message from "@model/base/message";
 
 export default interface CommandExecutionData {
   readonly trigger: string;
